+++ crates/request_editor/src/request_editor.rs
@@ -1950,91 +1950,6 @@ mod tests {
     }
 
     #[gpui::test]
-    async fn test_send_request_opens_response_panel(cx: &mut TestAppContext) {
-        cx.executor().allow_parking();
-
-        let temp_fs = TempFs::new(cx.executor());
-        let http_client = FakeHttpClient::with_response(StatusCode::NOT_FOUND);
-        let app_state =
-            cx.update(|cx| AppState::test_new(temp_fs.clone(), Some(http_client.clone()), cx));
-        let (tx, rx) = oneshot::channel();
-        let rx = Arc::new(Mutex::new(Some(rx)));
-
-        http_client.replace_handler({
-            move |_, request| {
-                assert_eq!(request.uri().path(), "/me");
-                let rx = rx.lock().take().unwrap();
-
-                async move { Ok(rx.await.unwrap()) }
-            }
-        });
-
-        init_test(app_state, cx);
-
-        temp_fs.insert_tree(
-            path!("project"),
-            json!({
-                "collection": {
-                    "request.toml": indoc! {r#"
-                        [meta]
-                        version = 1
-
-                        [http]
-                        method = "GET"
-                        url = "https://api.zaku.dev/me"
-                    "#}
-                }
-            }),
-        );
-
-        let project_path = temp_fs.path().join(path!("project"));
-        let project = Project::test_new(temp_fs.clone(), &project_path, cx).await;
-        let worktree_id = cx.update(|cx| project.read(cx).root_worktree(cx).unwrap().read(cx).id());
-        let (workspace, response_panel, cx) = build_workspace(&project, cx);
-
-        let request_path = ProjectPath {
-            worktree_id,
-            path: Arc::from(rel_path("collection/request.toml")),
-        };
-
-        workspace
-            .update_in(cx, |workspace, window, cx| {
-                workspace.open_path(request_path, None, true, window, cx)
-            })
-            .await
-            .unwrap()
-            .downcast::<RequestEditor>()
-            .unwrap();
-        workspace.update_in(cx, |workspace, window, cx| {
-            let focus_handle = workspace.focus_handle(cx);
-            window.focus(&focus_handle, cx);
-        });
-        cx.dispatch_action(actions::workspace::SendRequest);
-        cx.run_until_parked();
-
-        workspace.update_in(cx, |workspace, _, cx| {
-            let response_panel_id = Entity::entity_id(&response_panel);
-            let active_panel_id = workspace
-                .bottom_dock()
-                .read(cx)
-                .active_panel()
-                .map(|panel| panel.panel_id());
-
-            assert!(workspace.bottom_dock().read(cx).is_open());
-            assert_eq!(active_panel_id, Some(response_panel_id));
-        });
-
-        let response = Response::builder()
-            .status(StatusCode::OK)
-            .body(AsyncBody::from("response"))
-            .unwrap();
-        assert!(
-            matches!(tx.send(response).map_err(|_response| ()), Ok(())),
-            "response receiver should be active"
-        );
-    }
-
-    #[gpui::test]
     async fn test_send_request_respects_disabled(cx: &mut TestAppContext) {
         cx.executor().allow_parking();
 
@@ -2138,328 +2053,8 @@ mod tests {
             .body(AsyncBody::empty())
             .unwrap();
         assert!(
-            matches!(tx.send(response).map_err(|_response| ()), Ok(())),
+            matches!(tx.send(response), Ok(())),
             "response receiver should be active"
-        );
-    }
-
-    #[gpui::test]
-    async fn test_each_request_editor_has_its_own_response(cx: &mut TestAppContext) {
-        cx.executor().allow_parking();
-
-        let temp_fs = TempFs::new(cx.executor());
-        let http_client = FakeHttpClient::with_response(StatusCode::NOT_FOUND);
-        let app_state =
-            cx.update(|cx| AppState::test_new(temp_fs.clone(), Some(http_client.clone()), cx));
-        let (first_tx, first_rx) = oneshot::channel();
-        let (second_tx, second_rx) = oneshot::channel();
-        let first_rx = Arc::new(Mutex::new(Some(first_rx)));
-        let second_rx = Arc::new(Mutex::new(Some(second_rx)));
-        let first_response_delay = Duration::from_secs(5);
-        let second_response_delay = Duration::from_secs(3);
-        let executor = cx.executor();
-
-        http_client.replace_handler({
-            move |_, request| {
-                let (rx, response_delay) = match request.uri().path() {
-                    "/first" => (first_rx.lock().take().unwrap(), first_response_delay),
-                    "/second" => (second_rx.lock().take().unwrap(), second_response_delay),
-                    path => panic!("Unexpected request path: {path}"),
-                };
-                let executor = executor.clone();
-
-                async move {
-                    let response = rx.await.unwrap();
-                    executor.timer(response_delay).await;
-                    Ok(response)
-                }
-            }
-        });
-
-        init_test(app_state, cx);
-
-        temp_fs.insert_tree(
-            path!("project"),
-            json!({
-                "collection": {
-                    "first.toml": indoc! {r#"
-                        [meta]
-                        version = 1
-
-                        [http]
-                        method = "GET"
-                        url = "https://api.zaku.dev/first"
-                    "#},
-                    "second.toml": indoc! {r#"
-                        [meta]
-                        version = 1
-
-                        [http]
-                        method = "GET"
-                        url = "https://api.zaku.dev/second"
-                    "#}
-                }
-            }),
-        );
-
-        let project_path = temp_fs.path().join(path!("project"));
-        let project = Project::test_new(temp_fs.clone(), &project_path, cx).await;
-        let worktree_id = cx.update(|cx| project.read(cx).root_worktree(cx).unwrap().read(cx).id());
-        let (workspace, response_panel, cx) = build_workspace(&project, cx);
-        let pane = workspace.update_in(cx, |workspace, _, _| workspace.pane().clone());
-
-        let first_path = ProjectPath {
-            worktree_id,
-            path: Arc::from(rel_path("collection/first.toml")),
-        };
-        let second_path = ProjectPath {
-            worktree_id,
-            path: Arc::from(rel_path("collection/second.toml")),
-        };
-        let first_editor = workspace
-            .update_in(cx, |workspace, window, cx| {
-                workspace.open_path(first_path, None, true, window, cx)
-            })
-            .await
-            .unwrap()
-            .downcast::<RequestEditor>()
-            .unwrap();
-
-        pane.update_in(cx, |pane, window, cx| {
-            pane.send_request(window, cx);
-        });
-
-        let second_editor = workspace
-            .update_in(cx, |workspace, window, cx| {
-                workspace.open_path(second_path, None, true, window, cx)
-            })
-            .await
-            .unwrap()
-            .downcast::<RequestEditor>()
-            .unwrap();
-
-        pane.update_in(cx, |pane, window, cx| {
-            pane.send_request(window, cx);
-        });
-
-        let response = Response::builder()
-            .status(StatusCode::OK)
-            .body(AsyncBody::from("first response"))
-            .unwrap();
-        assert!(
-            matches!(first_tx.send(response).map_err(|_response| ()), Ok(())),
-            "response receiver should be active"
-        );
-
-        cx.executor().advance_clock(first_response_delay);
-        cx.run_until_parked();
-
-        assert_eq!(
-            response_panel.read_with(cx, |response_panel, cx| response_panel.text(cx)),
-            ""
-        );
-
-        let response = Response::builder()
-            .status(StatusCode::OK)
-            .body(AsyncBody::from("second response"))
-            .unwrap();
-        assert!(
-            matches!(second_tx.send(response).map_err(|_response| ()), Ok(())),
-            "response receiver should be active"
-        );
-
-        cx.executor().advance_clock(second_response_delay);
-        cx.run_until_parked();
-
-        assert_eq!(
-            first_editor.read_with(cx, |editor, cx| {
-                editor.response.as_ref().unwrap().read(cx).text(cx)
-            }),
-            "first response"
-        );
-        assert_eq!(
-            second_editor.read_with(cx, |editor, cx| {
-                editor.response.as_ref().unwrap().read(cx).text(cx)
-            }),
-            "second response"
-        );
-        assert_eq!(
-            response_panel.read_with(cx, |response_panel, cx| response_panel.text(cx)),
-            "second response"
-        );
-
-        let first_item_id = Entity::entity_id(&first_editor);
-        let item_index = pane.read_with(cx, |pane, _| {
-            pane.items()
-                .position(|item| item.item_id() == first_item_id)
-                .unwrap()
-        });
-
-        pane.update_in(cx, |pane, window, cx| {
-            pane.activate_item(item_index, true, false, window, cx);
-        });
-        cx.run_until_parked();
-
-        assert_eq!(
-            response_panel.read_with(cx, |response_panel, cx| response_panel.text(cx)),
-            "first response"
-        );
-    }
-
-    #[gpui::test]
-    async fn test_send_request_with_preview_request_editor(cx: &mut TestAppContext) {
-        cx.executor().allow_parking();
-
-        let temp_fs = TempFs::new(cx.executor());
-        let http_client = FakeHttpClient::with_response(StatusCode::NOT_FOUND);
-        let app_state =
-            cx.update(|cx| AppState::test_new(temp_fs.clone(), Some(http_client.clone()), cx));
-        let (first_tx, first_rx) = oneshot::channel();
-        let (second_tx, second_rx) = oneshot::channel();
-        let first_rx = Arc::new(Mutex::new(Some(first_rx)));
-        let second_rx = Arc::new(Mutex::new(Some(second_rx)));
-
-        http_client.replace_handler({
-            move |_, request| {
-                let rx = match request.uri().path() {
-                    "/first" => first_rx.lock().take().unwrap(),
-                    "/second" => second_rx.lock().take().unwrap(),
-                    path => panic!("Unexpected request path: {path}"),
-                };
-                async move { Ok(rx.await.unwrap()) }
-            }
-        });
-
-        init_test(app_state, cx);
-
-        temp_fs.insert_tree(
-            path!("project"),
-            json!({
-                "collection": {
-                    "first.toml": indoc! {r#"
-                        [meta]
-                        version = 1
-
-                        [http]
-                        method = "GET"
-                        url = "https://api.zaku.dev/first"
-                    "#},
-                    "second.toml": indoc! {r#"
-                        [meta]
-                        version = 1
-
-                        [http]
-                        method = "GET"
-                        url = "https://api.zaku.dev/second"
-                    "#}
-                }
-            }),
-        );
-
-        let project_path = temp_fs.path().join(path!("project"));
-        let project = Project::test_new(temp_fs.clone(), &project_path, cx).await;
-        let worktree_id = cx.update(|cx| project.read(cx).root_worktree(cx).unwrap().read(cx).id());
-        let (workspace, response_panel, cx) = build_workspace(&project, cx);
-        let pane = workspace.update_in(cx, |workspace, _, _| workspace.pane().clone());
-
-        let first_path = ProjectPath {
-            worktree_id,
-            path: Arc::from(rel_path("collection/first.toml")),
-        };
-        let second_path = ProjectPath {
-            worktree_id,
-            path: Arc::from(rel_path("collection/second.toml")),
-        };
-        let first_editor = workspace
-            .update_in(cx, |workspace, window, cx| {
-                workspace.open_path_preview(first_path, None, false, true, true, window, cx)
-            })
-            .await
-            .unwrap()
-            .downcast::<RequestEditor>()
-            .unwrap();
-
-        pane.update_in(cx, |pane, window, cx| {
-            pane.send_request(window, cx);
-        });
-        assert!(pane.read_with(cx, |pane, _| pane.preview_item_idx().is_none()));
-
-        let second_editor = workspace
-            .update_in(cx, |workspace, window, cx| {
-                workspace.open_path_preview(second_path, None, false, true, true, window, cx)
-            })
-            .await
-            .unwrap()
-            .downcast::<RequestEditor>()
-            .unwrap();
-
-        pane.update_in(cx, |pane, window, cx| {
-            pane.send_request(window, cx);
-        });
-
-        let first_item_id = Entity::entity_id(&first_editor);
-        assert!(pane.read_with(cx, |pane, _| {
-            pane.items().any(|item| item.item_id() == first_item_id)
-        }));
-
-        let response = Response::builder()
-            .status(StatusCode::OK)
-            .body(AsyncBody::from("first response"))
-            .unwrap();
-        assert!(
-            matches!(first_tx.send(response).map_err(|_response| ()), Ok(())),
-            "response receiver should be active"
-        );
-
-        cx.run_until_parked();
-
-        assert_eq!(
-            response_panel.read_with(cx, |response_panel, cx| response_panel.text(cx)),
-            ""
-        );
-
-        let response = Response::builder()
-            .status(StatusCode::OK)
-            .body(AsyncBody::from("second response"))
-            .unwrap();
-        assert!(
-            matches!(second_tx.send(response).map_err(|_response| ()), Ok(())),
-            "response receiver should be active"
-        );
-
-        cx.run_until_parked();
-
-        assert_eq!(
-            first_editor.read_with(cx, |editor, cx| {
-                editor.response.as_ref().unwrap().read(cx).text(cx)
-            }),
-            "first response"
-        );
-        assert_eq!(
-            second_editor.read_with(cx, |editor, cx| {
-                editor.response.as_ref().unwrap().read(cx).text(cx)
-            }),
-            "second response"
-        );
-        assert_eq!(
-            response_panel.read_with(cx, |response_panel, cx| response_panel.text(cx)),
-            "second response"
-        );
-
-        let item_index = pane.read_with(cx, |pane, _| {
-            pane.items()
-                .position(|item| item.item_id() == first_item_id)
-                .unwrap()
-        });
-
-        pane.update_in(cx, |pane, window, cx| {
-            pane.activate_item(item_index, true, false, window, cx);
-        });
-        cx.run_until_parked();
-
-        assert_eq!(
-            response_panel.read_with(cx, |response_panel, cx| response_panel.text(cx)),
-            "first response"
         );
     }
