+++ crates/zaku/Cargo.toml
@@ -81,6 +81,7 @@ gpui_platform = { workspace = true, features = [
 fs = { workspace = true, features = ["test"] }
 gpui = { workspace = true, features = ["test-support"] }
 http_client = { workspace = true, features = ["test"] }
+parking_lot = { workspace = true }
 path = { workspace = true, features = ["test"] }
 project = { workspace = true, features = ["test"] }
 response_panel = { workspace = true, features = ["test"] }
